python_stuff/requirements-python.txt
# Python dependencies for Imagen 4 testing
google-genai>=0.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0

